rename: move use query string hook
